Distribution.Nixpkgs.Meta: render full attr paths for maintainers
`with` has the sometimes surprising property of never shadowing an
existing binding in scope. Thus if we ever would add a maintainer that
has the same name as a package in scope, eval would break because a
derivation would be added to the maintainer list.

To prevent this we generate the list as `[ lib.maintainers.<maintainer> ]`
instead of `with lib.maintainers; [ <maintainer> ]`. For packages
with only a single maintainer, we even save space.

Resolves https://github.com/NixOS/cabal2nix/issues/499.