feat(Nixpkgs.Meta): indicate whether to render platforms via Maybe
Using `allKnownPlatforms` to determine if we need to render the
`platforms` or `hydraPlatforms` set is a bit awkward — it doesn't
correspond to lib.platforms.all (which is set by the Haskell generic
builder by default) and keeping it updated is probably an unnecessary
chore.

Thus we explicitly describe if the platforms attributes should be
generated by wrapping the respective platform set in a `Maybe`.