feat(Nixpkgs.Meta): add support for rendering badPlatforms
`badPlatforms` seems to be the cleanest way to support hackage2nix's
`unsupported-platforms` features without having to know a list of all
nixpkgs platforms `haskellPackages` supports statically in
distribution-nixpkgs. Since `stdenv.mkDerivation` will inspect both
`platforms` and `badPlatforms`, we can defer determining the effective
platform list to nix evaluation.

`badPlatforms` is an undocumented feature of nixpkgs at the moment and
some community members have suggested that it should be removed. This
is probably not an issue for us here though. If were to be removed, we
could still either emulate the feature in Haskell's generic builder or
emit a `subtractList` from `Distribution.Nixpkgs.Meta`.