feat(Distribution.Nixpkgs.Meta): support nixpkgs platform names
* Single (cabal) platforms are listed as they used to be

* Platform  groups are added to the list via Nix's `++` operator

The precise platform formatting is shown in the documentation comment
for `renderPlatforms` which is not exposed, but its REPL-examples are
checked by doctest. I'm personally quite happy with it. The only
annoyance is that nixpkgs platforms always start on a different line
than the binding even if there is no single platforms list. Fixing this
however would probably require an insufferable amount of screwing with
the pretty printer and make the code even more incomprehensible.