feat(Meta): implement parsing of NixpkgsPlatform from config format
nixpkgsPlatformFromString defines a simple string based format for
specifying NixpkgsPlatforms which is intended to be used by
hackage2nix's configuration file. It is based on the following mapping:

* lib.platform.* -> NixpkgsPlatformGroup
* platform.* -> NixpkgsPlatformGroup (as a shorthand)
* CPU-OS -> NixpkgsPlatformSingle (this is the current format)

This function is added to distribution-nixpkgs because it is easier to
implement, test and maintain there. It has the slight chance of being
confusing for users, but may also prove useful for someone else. Parsing
just Nix style system tuples into Cabal Platforms is (within limits)
supported by cabalPlatformFromSystem.