doc(Meta): warn about cabalPlatformFromSystem's differing parsing
Due to the reasons stated here, cabalPlatformSystem should always never
be used and it is probably a good idea to not expose it at all. I
originally thought we may want to use it in hackage2nix, but it also
only uses the Cabal Platform to evaluate Cabal files, so it should not
use this logic necessarily.