feat(Meta): no longer export cabalPlatformFromSystem
This function is not a general purpose conversion from a Nix system to a
Cabal Platform. We use Platform (for better or worse) to represent a Nix
system, not a Cabal platform. Thus nixpkgsPlatformFromString and the
wrapper type describe our intentions better. cabalPlatformFromSystem
should never be used to obtain a Platform and pass it to Cabal for some
purpose, so let's make that harder.

cabal2nix's Platform parsing can be independent from this and
implemented correctly: https://github.com/NixOS/cabal2nix/pull/555